services/sentiment-analytic/requirements.txt
absl-py==0.12.0
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
blis==0.4.1
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloud-tpu-client==0.10
cymem==2.0.3
decorator==4.4.2
dill==0.3.2
filelock==3.0.12
gast==0.3.3
google-api-core==1.28.0
google-api-python-client==1.8.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
gviz-api==1.9.0
h5py==2.10.0
httplib2==0.19.1
idna==2.9
importlib-metadata==1.6.0
isort==5.6.4
joblib==1.0.1
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
llvmlite==0.36.0
Markdown==3.3.4
mccabe==0.6.1
more-itertools==8.2.0
murmurhash==1.0.2
networkx==2.5
numba==0.53.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.0.3
pip-autoremove==0.9.1
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
protobuf==3.15.8
py==1.8.1
py4j==0.10.9
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfunctional==1.4.2
pylint==2.6.0
pyparsing==2.4.7
pyspark==3.0.1
pyspark-stubs==3.0.0.post1
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
regex==2021.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scipy==1.4.1
six==1.15.0
spacy==2.2.4
srsly==1.0.2
tabulate==0.8.7
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-profile==2.3.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
thinc==7.4.0
tokenizers==0.10.2
toml==0.10.1
tqdm==4.46.0
transformers==4.5.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
